rename : MockEntity 네이밍 변경(#231)
